Move a few things around
