+++ src/bin/src/packet_handlers/mod.rs
@@ -4,3 +4,6 @@ mod login_process;
 mod player;
 mod player_leave;
 mod tick_handler;
+
+mod velocity;
+mod whitelist;
